# backend/docs/swagger.yaml
basePath: /
definitions:
...
  balance_handlers.Response:
    properties:
      data: {}
      status:
        type: string
      statusCode:
        type: integer
    type: object
...
host: localhost:8000
info:
  contact:
    email: allan.cordeiro.santos@gmail.com
    name: Allan Cordeiro
    url: http://www.allancordeiro.com
  description: Product API with stock management
  termsOfService: http://www.swagger.io/terms
  title: Despensa Faculdade Impacta's Project
  version: "1.0"
paths:
  /api/products/{productID}/balance:
    get:
      consumes:
      - application/json
      description: Get product history of reductions
      parameters:
      - description: product identifier
        in: path
        name: productID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/balance_handlers.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/balance_handlers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/balance_handlers.Response'
      summary: Get Product Balance
      tags:
      - products
...
swagger: "2.0"